interests-map/requirements.txt
werkzeug
gunicorn
seaborn
scipy
numpy
scikit-learn
umap-learn
matplotlib
collections-extended
flask
requests
sentence-transformers